requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
boto3==1.29.7
Werkzeug==2.3.7
pytest==7.4.2
